requirements.txt
openai
pyyaml
flask
uuid
flask_cors
